Fix for non-owner can delete/edit the image issue

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -102,9 +102,10 @@ public class ImageController {
             String error = "Only the owner of the image can edit the image";
             model.addAttribute("editError", error);
             return "images/image";
-        }
+        } else {
 
-        return "images/edit";
+            return "images/edit";
+        }
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -146,9 +147,19 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId,Model model, HttpSession session) {
+        Image image = imageService.getImage(imageId);
+        User user = (User) session.getAttribute("loggeduser");
+        if(user.getId()!= image.getUser().getId()) {
+            String error = "Only the owner of the image can delete the image";
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("deleteError", error);
+            model.addAttribute("image", image);
+            model.addAttribute("tags", tags);
+            return "images/image";
+        } else {
+            imageService.deleteImage(imageId);
+        return "redirect:/images";}
     }
